Added "FlxState.backgroundColor", "FlxState.useMouse", and "FlxState.isSubState"
The sub-state no longer has a "child" displaying the background color,
but the background color is instead passed directly to the camera.

If a "parent state" uses the mouse cursor, and the sub state does not,
when the sub state closes, the parent will go back to displaying a mouse
cursor (and vice versa).

The "isSubState" option allows FlxSubStates to act just like FlxStates
(which may be useful for when the "pause" menu is identical to the "main
menu".